web/requirements.txt
fastapi==0.115.12
jinja2==3.1.6
langchain==0.3.23
langchain_neo4j==0.4.0
langchain_openai==0.3.12
pydantic==2.11.3
pydantic_settings==2.8.1
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.34.1
numpy==2.2.4